Fix: Rename `install.php` to `install.php.disabled` since config.php is mounted into the container
In a dockerized setup, `config.php` is mounted in the `bf2sclone` container with write permissions since as of `2.2.0`, `bf2sclone` expects `config.inc.php` to be writeable.  If `install.php` is present, any user can write to the `config.inc.php`, which shouldn't happen.

Now `install.php` is renamed to `install.php.disabled` so that it is not passed to `php`.